backend/requirements.txt
# Python backend dependencies for Reasoning System

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Azure SDK
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
openai==1.7.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Validation & Serialization
email-validator==2.1.0
python-dateutil==2.8.2

# Celery for background tasks
celery==5.3.4
kombu==5.3.4

# Monitoring & Logging
sentry-sdk==1.39.2
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
click==8.1.7
tenacity==8.2.3

# Code Analysis & Graph Generation
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.2
tree-sitter-typescript==0.21.1
networkx==3.2.1
neo4j==5.15.0

# Docker SDK for Sandbox Execution
docker==7.0.0

# AI/LLM APIs
anthropic==0.8.1
